+++ tests/testthat/test-handle.R
@@ -389,6 +389,81 @@ describe("Router", {
     })
   })
 
+  describe("req$baseUrl", {
+    it("should contain the stripped path", {
+      router <- Router$new()
+      router$use("/foo", sawBase)
+
+      server <- createServer(router)
+      r <- fetch(server, "/foo/bar")
+
+      expect_identical(
+        list(
+          r$status,
+          r$body
+        ),
+        list(
+          200L,
+          "saw /foo"
+        )
+      )
+    })
+
+    it("should contain the stripped path from multiple levels", {
+      router1 <- Router$new()
+      router2 <- Router$new()
+
+      router1$use("/foo", router2)
+      router2$use("/bar", sawBase)
+
+      server <- createServer(router1)
+      r <- fetch(server, "/foo/bar/baz")
+
+      expect_identical(
+        list(
+          r$status,
+          r$body
+        ),
+        list(
+          200L,
+          "saw /foo/bar"
+        )
+      )
+    })
+
+    it("should be altered correctly", {
+      router <- Router$new()
+      sub1 <- Router$new()
+      sub2 <- Router$new()
+      sub3 <- Router$new()
+
+      sub3$get("/zed", setsawBase(1))
+
+      sub2$use("/baz", sub3)
+
+      sub1$use("/", setsawBase(2))
+
+      sub1$use("/bar", sub2)
+      sub1$use("/bar", setsawBase(3))
+
+      router$use(setsawBase(4))
+      router$use("/foo", sub1)
+      router$use(setsawBase(5))
+      router$use(hello_world)
+
+      server <- createServer(router)
+
+      r <- fetch(server, "/foo/bar/baz/zed")
+
+      expect_identical(r$status, 200L)
+      expect_identical(r$headers$`x-saw-base1`, "/foo/bar/baz")
+      expect_identical(r$headers$`x-saw-base2`, "/foo")
+      expect_identical(r$headers$`x-saw-base3`, "/foo/bar")
+      expect_identical(r$headers$`x-saw-base4`, "")
+      expect_identical(r$headers$`x-saw-base5`, "")
+    })
+  })
+
   describe("req$PATH_INFO", {
     it("should strip path from req$PATH_INFO", {
       router <- Router$new()
@@ -511,6 +586,130 @@ describe("Router", {
       should_hit_handle(r, 4)
       expect_equal(r$status, 200L)
       expect_equal(r$body, "saw GET /bar")
+    })
+  })
+
+  describe("static paths", {
+    css_content <- raw_file_content(test_path("static/css/main.css"))
+    html_content <- raw_file_content(test_path("static/index.html"))
+    it("should serve files", {
+      router <- Router$new()
+      router$static(
+        test_path("static"),
+        "/"
+      )
+
+      server <- createServer(router)
+
+      r <- fetch(server, "/")
+
+      expect_identical(r$status, 200L)
+      expect_identical(charToRaw(r$body), html_content)
+
+      r <- fetch(server, "/css/main.css")
+
+      expect_identical(r$status, 200L)
+      expect_identical(charToRaw(r$body), css_content)
+    })
+
+    describe("should respect staticPath arguments", {
+      it("sholuld respect indexhtml argument", {
+        router <- Router$new()
+        router$static(
+          test_path("static"),
+          "/",
+          indexhtml = FALSE
+        )
+
+        server <- createServer(router)
+
+        r <- fetch(server, "/")
+
+        expect_identical(r$status, 404L)
+        expect_identical(r$body, "404 Not Found\n")
+      })
+
+      it("should respect fallthrough argument", {
+        router <- Router$new()
+        router$static(
+          test_path("static"),
+          "/",
+          fallthrough = TRUE
+        )
+
+        server <- createServer(router)
+
+        # Should go through our finalHandler (slower, not advised)
+        r <- fetch(server, "/foo")
+        expect_identical(r$status, 404L)
+        expect_true(grepl(pattern = "Cannot GET /foo", x = r$body))
+      })
+
+      it("should respect html_charset argument", {
+        router <- Router$new()
+        router$static(
+          test_path("static"),
+          "/",
+          html_charset = ""
+        )
+
+        server <- createServer(router)
+
+        r <- fetch(server, "/")
+        expect_identical(r$status, 200L)
+        expect_identical(r$headers$`content-type`, "text/html")
+
+        # Default behaviour
+        router1 <- Router$new()
+        router1$static(
+          test_path("static"),
+          "/"
+        )
+
+        server <- createServer(router1)
+
+        r <- fetch(server, "/")
+        expect_identical(r$status, 200L)
+        expect_identical(r$headers$`content-type`, "text/html; charset=utf-8")
+      })
+
+      it("should respect headers argument", {
+        router <- Router$new()
+        router$static(
+          test_path("static"),
+          "/",
+          headers = list(
+            "X-Powered-By" = "routing"
+          )
+        )
+
+        server <- createServer(router)
+
+        r <- fetch(server, "/")
+        expect_identical(r$status, 200L)
+        expect_identical(charToRaw(r$body), html_content)
+        expect_identical(r$headers$`x-powered-by`, "routing")
+      })
+      it("should respect validation argument", {
+        router <- Router$new()
+        router$static(
+          test_path("static"),
+          "/",
+          validation = c('"foo" == "bar"')
+        )
+
+        server <- createServer(router)
+
+        r <- fetch(server, "/", headers = list(foo = "zoo"))
+
+        expect_identical(r$status, 403L)
+        expect_identical(r$body, "403 Forbidden\n")
+
+        r <- fetch(server, "/", headers = list(foo = "bar"))
+
+        expect_identical(r$status, 200L)
+        expect_identical(charToRaw(r$body), html_content)
+      })
     })
   })
 })